Smart contract feature update

+++ include/class/SmartContractEngine.php
@@ -11,51 +11,13 @@ class SmartContractEngine
 		return ROOT . "/tmp/sc";
 	}
 
-	static function storeState($sc_address, $height, $state) {
-
-		if(self::$virtual) {
-			$state_file = self::getRunFolder() . "/{$sc_address}_state.json";
-			file_put_contents($state_file, json_encode($state));
-		} else {
-			foreach ($state as $name => $value) {
-				if(is_array($value)) {
-					foreach ($value as $var_key => $var_value) {
-						self::insertSmartConractStateRow($sc_address, $name, $var_value, $var_key, $height);
-					}
-				} else {
-					self::insertSmartConractStateRow($sc_address, $name, $value, null, $height);
-				}
-			}
-		}
-
-
-	}
-
-	private static function insertSmartConractStateRow($sc_address,$name,$value,$key,$height) {
-		global $db;
-
-		if(strlen($value) > 1000) {
-			throw new Exception("Storing value for variable $name key $key exceeds 1000 characters");
-		}
-
-		$sql="replace into smart_contract_state (sc_address, variable, var_key, var_value, height)
-					values (:sc_address, :variable, :var_key, :var_value, :height)";
-		$bind = [
-			":sc_address"=>$sc_address,
-			":variable"=>$name,
-			":var_key"=>$key,
-			":var_value"=>"$value",
-			":height"=>$height
-		];
-		$res = $db->run($sql, $bind);
-		if($res === false) {
-			throw new Exception("Error storing variable $name key=$key for Smart Contract $sc_address: ".$db->errorInfo()[2]);
-		}
-	}
 
 	static function verifySmartContract($sc_address, $test = false) {
 
 		if(self::$virtual) {
+            if(self::$smartContract['address']!=$sc_address) {
+                throw new Exception("Not found Smart Contract with address $sc_address");
+            }
 			return self::$smartContract;
 		}
 
@@ -87,7 +49,7 @@ class SmartContractEngine
 		return $smartContract;
 	}
 
-	static function SCGet($sc_address, $property, $key=null, &$error=null) {
+	static function get($sc_address, $property, $key=null, &$error=null) {
 		return try_catch(function () use ($sc_address, $property, $key, &$error){
 
 			$smartContract = SmartContractEngine::verifySmartContract($sc_address);
@@ -110,13 +72,11 @@ class SmartContractEngine
 		}, $error);
 	}
 
-	static function exec($transaction, $method, $height=0, $params=[], &$error=null) {
+	static function exec($transaction, $method, $height=0, $params=[], &$error=null, $test=false) {
 		$sc_address = $transaction->dst;
-		return try_catch(function () use ($sc_address, $transaction, $method, $params, $height, $error) {
+		return try_catch(function () use ($sc_address, $transaction, $method, $params, $height, $error,$test) {
 
 			$sc_exec_file = SmartContractEngine::buildRunCode($sc_address);
-
-			SmartContractEngine::buildState($sc_address);
 
 			if(!empty($params) && !is_array($params)) {
 				$params = [$params];
@@ -127,19 +87,17 @@ class SmartContractEngine
 				'method' => $method,
 				'params' => $params,
 				'transaction' => $transaction,
-				'height'=>$height
+				'height'=>$height,
+                "test"=>$test,
+                "virtual"=>self::$virtual
 			];
 
 			$cmd_args = base64_encode(json_encode($cmd_args));
 
 			$cmd = "$sc_exec_file $cmd_args";
 
-			$output = self::isolateCmd($cmd);
-
-			$out = self::processOutput($output);
-			$state = $out['state'];
-
-			SmartContractEngine::storeState($sc_address, $height, $state);
+            $res = self::isolateCmd($cmd);
+			self::processOutput($res);
 
 			return true;
 
@@ -147,42 +105,39 @@ class SmartContractEngine
 
 	}
 
-	static function call($sc_address, $method, $params=[], &$error=null) {
+	static function view($sc_address, $method, $params=[], &$error=null) {
 		return try_catch(function () use ($sc_address, $method, $params) {
 
 			$sc_exec_file = SmartContractEngine::buildRunCode($sc_address);
-
-			SmartContractEngine::buildState($sc_address);
 
 			if(!empty($params) && !is_array($params)) {
 				$params = [$params];
 			}
 
 			$cmd_args = [
-				'type'=>'call',
+				'type'=>'view',
 				'method' => $method,
 				'params' => $params,
+                "virtual"=>self::$virtual
 			];
 
 			$cmd_args = base64_encode(json_encode($cmd_args));
 
 			$cmd = "$sc_exec_file $cmd_args";
 
-			$output = self::isolateCmd($cmd);
-			$out = self::processOutput($output);
+			$res = self::isolateCmd($cmd);
+			$out = self::processOutput($res);
 			$response = $out['response'];
 			return $response;
 
 		}, $error);
 	}
 
-	static function send($transaction, $method, $height=0, $params=[], &$error=null) {
+	static function send($transaction, $method, $height=0, $params=[], &$error=null, $test=false) {
 		$sc_address = $transaction->src;
-		return try_catch(function () use ($sc_address, $transaction, $method, $params, $height, $error) {
+		return try_catch(function () use ($sc_address, $transaction, $method, $params, $height, $test) {
 
 			$sc_exec_file = SmartContractEngine::buildRunCode($sc_address);
-
-			SmartContractEngine::buildState($sc_address, $height);
 
 			if(!empty($params) && !is_array($params)) {
 				$params = [$params];
@@ -193,20 +148,18 @@ class SmartContractEngine
 				'method' => $method,
 				'params' => $params,
 				'transaction' => $transaction,
-				'height'=>$height
+				'height'=>$height,
+                "test"=>$test,
+                "virtual"=>self::$virtual
 			];
 
 			$cmd_args = base64_encode(json_encode($cmd_args));
 
 			$cmd = "$sc_exec_file $cmd_args";
 
-			$output = self::isolateCmd($cmd);
+            $res = self::isolateCmd($cmd);
+			self::processOutput($res);
 
-			$out = self::processOutput($output);
-
-			$state = $out['state'];
-
-			SmartContractEngine::storeState($sc_address, $height, $state);
 
 			return true;
 
@@ -215,26 +168,65 @@ class SmartContractEngine
 	}
 
 	static function isolateCmd($cmd) {
-		$exec_cmd = "php ".XDEBUG_CLI." -d disable_functions=exec,passthru,shell_exec,system,proc_open,popen,curl_exec,curl_multi_exec,parse_ini_file,show_source,ini_set ";
-		$exec_cmd.= " -d memory_limit=".SC_MEMORY_LIMIT;
-		$exec_cmd.= " -d open_basedir=".self::getRunFolder();
+        $debug="-dxdebug.start_with_request=1";
+        $debug="";
+
+
+        $allowed_files = [
+            ROOT . "/chain_id",
+            ROOT . "/include/sc.inc.php",
+            ROOT . "/include/db.inc.php",
+            ROOT . "/include/common.functions.php",
+            ROOT . "/include/class/sc"
+        ];
+
+        if(file_exists(ROOT."/chain_id")) {
+            $chain_id = trim(file_get_contents(ROOT."/chain_id"));
+            $allowed_files[]=ROOT . "/include/coinspec.".$chain_id.".inc.php";
+        }
+
+        $allowed_files_list = implode(":", $allowed_files);
+
+        global $_config;
+        $config=[
+            "db_connect"=>$_config['db_connect'],
+            "db_user"=>$_config['db_user'],
+            "db_pass"=>$_config['db_pass'],
+        ];
+
+        $config=base64_encode(json_encode($config));
+        $error_reporting=E_ALL^E_NOTICE;
+
+		$exec_cmd = "CONFIG=$config php $debug -d disable_functions=exec,passthru,shell_exec,system,proc_open,popen,curl_exec,curl_multi_exec,parse_ini_file,show_source,ini_set,getenv,sleep,set_time_limit,error_reporting ";
+		$exec_cmd.= " -d memory_limit=".SC_MEMORY_LIMIT." -d max_execution_time=".SC_MAX_EXEC_TIME." -d error_reporting=$error_reporting";
+		$exec_cmd.= " -d open_basedir=".self::getRunFolder().":".$allowed_files_list;
 		$exec_cmd.= " -f $cmd ";
 		$exec_cmd.= " 2>&1";
 		$output = shell_exec ($exec_cmd);
 		$lines2 = [];
 		$lines = explode(PHP_EOL, $output);
+        $errors=[];
 		foreach($lines as $line) {
 			if(strpos($line, 'PHP Startup:')===0) {
+                $errors[]=trim($line);
 				continue;
 			}
 			if(strpos($line, 'PHP Warning:')===0) {
+                $errors[]=trim($line);
+				continue;
+			}
+			if(strpos($line, 'PHP Fatal error:')===0) {
+                $errors[]=trim($line);
 				continue;
 			}
 			$lines2[]=$line;
 		}
 		$output = implode(PHP_EOL, $lines2);
 		$output = trim($output);
-		return $output;
+		return [
+            "output"=>$output,
+            "errors"=>$errors
+        ];
 	}
 
 	static function buildRunCode($sc_address, $test=false) {
@@ -244,40 +236,43 @@ class SmartContractEngine
 		$data = json_decode(base64_decode($code), true);
 		$code = base64_decode($data['code']);
 
-		$sc_dir = self::getRunFolder();
-		if(!file_exists($sc_dir)) {
-			@mkdir($sc_dir);
-		}
+		return self::buildRunFile($sc_address, $code);
+	}
 
-		$phar_file = $sc_dir . "/$sc_address.phar";
-		if(!file_exists($phar_file) || DEVELOPMENT) {
-			file_put_contents($phar_file, $code);
-			@chmod($phar_file, 0777);
-		}
+    static function buildDeployCode(Transaction  $transaction, $test=false) {
 
-		$sc_run_file = $sc_dir. "/{$sc_address}_run.php";
-		if(file_exists($sc_run_file) && !DEVELOPMENT) {
-			return $sc_run_file;
-		}
+        $sc_address=$transaction->dst;
+        $data = json_decode(base64_decode($transaction->data), true);
+        $code = base64_decode($data['code']);
 
-		self::copyRequiredFiles(
-			[
-				"/include/class/SmartContractWrapper.php" => "/SmartContractWrapper.php",
-				"/include/class/SmartContractBase.php" => "/SmartContractBase.php",
-				"/include/common.functions.php" => "/common.functions.php"
-			]
-		);
+        return self::buildRunFile($sc_address, $code);
+    }
 
-		$run_code = "<?php
+    static function buildRunFile($sc_address, $code) {
+        $sc_dir = self::getRunFolder();
+        if(!file_exists($sc_dir)) {
+            @mkdir($sc_dir);
+        }
 
-		set_time_limit(".SC_MAX_EXEC_TIME.");
-		
-		error_reporting(E_ALL^E_NOTICE);
-		define(\"ROOT\", __DIR__);
+        $phar_file = $sc_dir . "/$sc_address.phar";
+        if(!file_exists($phar_file) || DEVELOPMENT) {
+            file_put_contents($phar_file, $code);
+            @chmod($phar_file, 0777);
+        }
+
+        $sc_run_file = $sc_dir. "/{$sc_address}_run.php";
+        if(file_exists($sc_run_file) && !DEVELOPMENT) {
+            return $sc_run_file;
+        }
+
+        $run_code = "<?php
+
+        if(function_exists('xdebug_disable')) {
+            xdebug_disable();
+        }
 		define(\"SC_ADDRESS\", \"$sc_address\");
 		
-		require_once 'common.functions.php';
-		require_once 'SmartContractWrapper.php';
+		require_once dirname(dirname(__DIR__)) . '/include/sc.inc.php';
 		require_once \"phar://$phar_file\";
 		
 		ob_start();
@@ -287,58 +282,40 @@ class SmartContractEngine
 		
 		";
 
-		$sc_run_file = $sc_dir. "/{$sc_address}_run.php";
-		file_put_contents($sc_run_file, $run_code);
+        $sc_run_file = $sc_dir. "/{$sc_address}_run.php";
+        file_put_contents($sc_run_file, $run_code);
 
-		return $sc_run_file;
-	}
-
-	static function copyRequiredFiles($files) {
-		foreach($files as $src=>$dst) {
-			$file_src = ROOT . $src;
-			$file_dst = self::getRunFolder() . $dst;
-			if(!file_exists($file_dst) ) {
-				copy($file_src, $file_dst);
-			} else {
-				if(md5_file($file_src) != md5_file($file_dst)) {
-					copy($file_src, $file_dst);
-				}
-			}
-		}
-	}
-
-	static function buildState($sc_address) {
-		$state = self::loadState($sc_address);
-		$state_file = self::getRunFolder() . "/{$sc_address}_state.json";
-		@file_put_contents($state_file, json_encode($state));
-	}
-
+        return $sc_run_file;
+    }
 
 	static function loadState($sc_address) {
 		global $db;
 
 		if(self::$virtual) {
 			$state = [];
-			$state_file = self::getRunFolder() . "/{$sc_address}_state.json";
+			$state_file = self::getRunFolder() . "/{$sc_address}.state.json";
 			if(file_exists($state_file)) {
 				$state = file_get_contents($state_file);
 				$state = json_decode($state, true);
 			}
 		} else {
-			$sql="select max(height) from smart_contract_state where sc_address = :address";
-			$height = $db->single($sql, [":address"=> $sc_address]);
-			$state = [];
-			if($height) {
-				$sql = "select * from smart_contract_state where sc_address = :address and height = :height";
-				$rows = $db->run($sql, [":address"=>$sc_address, ":height" => $height]);
-				foreach ($rows as $row) {
-					if($row['var_key']!==null) {
-						$state[$row['variable']][$row['var_key']]=$row['var_value'];
-					} else {
-						$state[$row['variable']]=$row['var_value'];
-					}
-				}
-			}
+            $state = [];
+            $sql="select ss.variable, ss.var_key, ss.var_value
+            from (select s.sc_address, s.variable, ifnull(s.var_key, 'null') as var_key, max(s.height) as height
+                  from smart_contract_state s
+                  where s.sc_address = :address
+                  group by s.variable, s.var_key) as last_vars
+                     join smart_contract_state ss on (ss.sc_address = last_vars.sc_address and ss.variable = last_vars.variable
+                and ifnull(ss.var_key, 'null') = last_vars.var_key and ss.height = last_vars.height);
+            ";
+            $rows = $db->run($sql, [":address"=> $sc_address]);
+            foreach ($rows as $row) {
+                if($row['var_key']!==null) {
+                    $state[$row['variable']][$row['var_key']]=$row['var_value'];
+                } else {
+                    $state[$row['variable']]=$row['var_value'];
+                }
+            }
 		}
 
 		return $state;
@@ -353,23 +330,15 @@ class SmartContractEngine
 
 			$phar_file = self::getRunFolder() . "/$name.phar";
 
-			file_put_contents($phar_file, $code);
+			$res = file_put_contents($phar_file, $code);
+            if(!$res) {
+                throw new Error("Unable to write phar file $phar_file");
+            }
 
-			self::copyRequiredFiles(
-				[
-					"/include/class/SmartContractWrapper.php" => "/SmartContractWrapper.php",
-					"/include/class/SmartContractBase.php" => "/SmartContractBase.php",
-					"/include/common.functions.php" => "/common.functions.php"
-				]
-			);
+			$sc_verify_file = self::getRunFolder() . "/{$name}_verify.php";
+			$sc_verify_content="<?php
 
-			$sc_test_file = self::getRunFolder() . "/{$name}_verify.php";
-			$sc_test_content="<?php
-
-			set_time_limit(".SC_MAX_EXEC_TIME.");
-
-			require_once 'common.functions.php';
-			require_once 'SmartContractWrapper.php';
+			require_once dirname(dirname(__DIR__)) . '/include/sc.inc.php';
 
 			ob_start();
 			require_once '$phar_file';
@@ -377,15 +346,17 @@ class SmartContractEngine
 			\$smartContractWrapper->verify();
 			";
 
-			file_put_contents($sc_test_file, $sc_test_content);
+			$res = file_put_contents($sc_verify_file, $sc_verify_content);
+            if(!$res) {
+                throw new Error("Unable to write verify phar file $sc_verify_file");
+            }
 
-			$cmd = "$sc_test_file";
-			$output = self::isolateCmd($cmd);
-
-			self::processOutput($output);
+			$cmd = "$sc_verify_file";
+            $res = self::isolateCmd($cmd);
+			self::processOutput($res);
 
 			if(!DEVELOPMENT) {
-				unlink($sc_test_file);
+				unlink($sc_verify_file);
 			}
 
 			return true;
@@ -393,14 +364,17 @@ class SmartContractEngine
 		}, $error);
 	}
 
-	static function processOutput($output) {
+	static function processOutput($res) {
+        $output=$res['output'];
+        $errors=$res['errors'];
 		$output_decoded = json_decode($output , true);
 
 		if(!is_array($output_decoded)) {
-			throw new Exception("Smart contract failed: $output");
+			throw new Exception("Smart contract failed: $output errors=".implode(PHP_EOL, $errors));
 		} else {
 			if($output_decoded['status']=='error') {
-				throw new Exception("Smart contract failed: ".$output_decoded['error']);
+				throw new Exception("Smart contract failed: ".$output_decoded['error']. " errors=".implode(PHP_EOL, $errors) .
+                " trace=".$output_decoded['trace']);
 			}
 		}
 		return $output_decoded['data'];
@@ -417,12 +391,12 @@ class SmartContractEngine
 			$cmd_args = base64_encode(json_encode($cmd_args));
 
 			$cmd = "$sc_exec_file $cmd_args";
-			$output = self::isolateCmd($cmd);
-			$out = self::processOutput($output);
+            $res = self::isolateCmd($cmd);
+			$out = self::processOutput($res);
 			if(!self::$virtual) {
 //				unlink($sc_exec_file);
 			}
-			return $out['interface'];
+            return $out['interface'];
 		}, $error);
 	}
 
@@ -432,10 +406,12 @@ class SmartContractEngine
 		$sc_address = $transaction->dst;
 		return try_catch(function () use ($sc_address, $height, $test, $transaction) {
 
-			$sc_exec_file = SmartContractEngine::buildRunCode($sc_address, $test);
+			$sc_exec_file = SmartContractEngine::buildDeployCode($transaction, $test);
 
-			$state_file = self::getRunFolder() . "/{$sc_address}_state.json";
-			@unlink($state_file);
+            if(self::$virtual) {
+                $state_file = ROOT . '/tmp/sc/'.$sc_address.'.state.json';
+                @unlink($state_file);
+            }
 
 			$data = $transaction->data;
 			$data = json_decode(base64_decode($data), true);
@@ -443,7 +419,9 @@ class SmartContractEngine
 			$cmd_args = [
 				'type'=>'deploy',
 				'transaction' => $transaction,
-				'height'=>$height
+				'height'=>$height,
+                "test"=>$test,
+                "virtual"=>self::$virtual
 			];
 
 			if(isset($data['params'])) {
@@ -453,14 +431,12 @@ class SmartContractEngine
 			$cmd_args = base64_encode(json_encode($cmd_args));
 
 			$cmd = "$sc_exec_file $cmd_args";
-			$output = self::isolateCmd($cmd);
-			$out = self::processOutput($output);
-			$state = $out['state'];
-			SmartContractEngine::storeState($sc_address, $height, $state);
-			if(!self::$virtual) {
+            $res = self::isolateCmd($cmd);
+			self::processOutput($res);
+			if(!self::$virtual && !DEVELOPMENT) {
 				unlink($sc_exec_file);
 			}
-			return $output;
+			return $res['output'];
 		}, $error);
